+++ src/main/java/com/thientvse/icommerce/services/impl/ProductServiceImpl.java
@@ -41,7 +41,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Product getDetailProduct(int id) {
+    public Product getDetailProduct(long id) {
         return productRepository.findById(id).orElseThrow(ObjectNotFoundException::new);
     }
 }
